refactor(terminal): 拆出查找/右键链接菜单与外壳组件，Terminal.tsx 收敛为组合层
- 新增 components/terminal/：useTerminalSearch（缓冲区查找与 Ctrl+F 唤起）、useTerminalMenus（右键/链接菜单动作与外点关闭）、TerminalSearchBar（内容查找栏）、TerminalMenus（右键菜单/链接菜单/链接 toast）、TerminalSurface（色调壁纸层、状态栏、xterm 视口侧栏）
- Terminal.tsx 999 -> 339 行（仅保留共享状态与 hook/组件组合）
